requirements.txt
apache-beam[gcp]==2.68.0
pyarrow==18.1.0
google-cloud-bigquery==3.38.0
google-cloud-secret-manager==2.24.0
PyYAML==6.0.3
Jinja2==3.1.6